updated gendocs script

+++ scripts/gendocs.ps1
@@ -1,7 +1,7 @@
 Push-Location
 Set-Location "$PSScriptRoot\..\docs"
-Remove-Item .\_source\source\*
-Remove-Item -Recurse .\_build\apidoc\*
+if (Test-Path ".\_source\source\") { Remove-Item .\_source\source\* }
+if (Test-Path ".\_build\apidoc\*") { Remove-Item -Recurse .\_build\apidoc\* }
 sphinx-apidoc --implicit-namespaces -M -T -o .\_source\source\ ..\src\netdox\
 sphinx-build -Ea -b psml .\_source\ .\_build\apidoc
 Pop-Location
